requirements.txt
cryptography
aiomysql
fastapi
tortoise-orm
tortoise
marshmallow
python-dotenv
selenium
beautifulsoup4
pytest
uvicorn
asgi_lifespan
pyjwt